+++ Sketch/libraries/CNCLibEx/src/U8GLcd_Menu.cpp
@@ -171,10 +171,10 @@ bool CU8GLcd::DrawLoopMenu(EnumAsByte(EDrawLoopType) type, uintptr_t data)
 	{
 		return true;
 	}
-	//	if (type==DrawLoopQueryTimerout)	{ *((unsigned long*)data) = 2000; return true; }
+	//	if (type==DrawLoopQueryTimerout)	{ *((uint32_t*)data) = 2000; return true; }
 	if (type == DrawLoopQueryTimerout)
 	{
-		*((unsigned long*)data) = 250;
+		*((uint32_t*)data) = 250;
 		return true;
 	}
 	if (type != DrawLoopDraw)
