+++ Sketch/libraries/StepperLib/src/ReadPinIOTriggerControl.h
@@ -21,7 +21,7 @@
 
 ////////////////////////////////////////////////////////
 
-template <pin_t PIN, uint8_t ONVALUE, unsigned long STABLETIME>
+template <pin_t PIN, uint8_t ONVALUE, uint32_t STABLETIME>
 class CReadPinIOTriggerControl
 {
 public:
@@ -35,7 +35,7 @@ public:
 	{
 		if (CHAL::digitalRead(PIN) == ONVALUE)
 		{
-			unsigned long now = millis();
+			uint32_t now = millis();
 
 			if (_timeOn == 0) // first on
 			{
@@ -51,7 +51,7 @@ public:
 
 private:
 
-	unsigned long _timeOn = 0;
+	uint32_t _timeOn = 0;
 };
 
 ////////////////////////////////////////////////////////
